% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pcnt_games_won.R
\name{pcnt_games_won}
\alias{pcnt_games_won}
\title{Percentage of Games Won}
\usage{
pcnt_games_won(data_match, team_code)
}
\arguments{
\item{data_match}{The data about matches.}

\item{team_code}{The code of the team to calculate result for.}
}
\value{
The percentage of games won.
}
\description{
Percentage of Games Won
}
