% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_pcnt_successful_free_throws_match.R
\name{pcnt_successful_free_throws_match}
\alias{pcnt_successful_free_throws_match}
\title{Percentage Successful Free Throws Per Match}
\usage{
pcnt_successful_free_throws_match(total_free_throws, successful_free_throws)
}
\arguments{
\item{total_free_throws}{A vector of total free throws per match.}

\item{successful_free_throws}{A vector of successful free throws per match.}
}
\value{
The percentage of successful free throws per match.
}
\description{
Percentage Successful Free Throws Per Match
}
